Refactoring runner.py - certificates handling
